TASK: Apply optimization for high risk components
